+++ cluster-operator/src/main/java/io/strimzi/operator/cluster/model/AbstractModel.java
@@ -1120,21 +1120,25 @@ public abstract class AbstractModel {
         return KafkaResources.clusterCaKeySecretName(cluster);
     }
 
-    protected static Map<String, String> mergeAnnotations(Map<String, String> internal, Map<String, String> template) {
+    protected static Map<String, String> mergeAnnotations(Map<String, String> internal, Map<String, String>... templates) {
         Map<String, String> merged = new HashMap<>();
 
         if (internal != null) {
             merged.putAll(internal);
         }
 
-        if (template != null) {
-            for (String key : template.keySet()) {
-                if (key.contains("strimzi.io")) {
-                    throw new IllegalArgumentException("User annotations includes a Strimzi annotation: " + key);
+        if (templates != null) {
+            for (Map<String, String> template : templates) {
+                if (template != null) {
+                    for (String key : template.keySet()) {
+                        if (key.contains("strimzi.io")) {
+                            throw new InvalidResourceException("User annotations includes a Strimzi annotation: " + key);
+                        }
+                    }
+
+                    merged.putAll(template);
                 }
             }
-
-            merged.putAll(template);
         }
 
         return merged;