+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Reduce the number of images needed to run Strimzi to just two: `kafka` and `operator`.
 * Add support for unprivileged users to install the operator with Helm
 * Support for Kafka 2.2.0
+* Support off-cluster access using Kubernetes Nginx Ingress
 
 ## 0.11.0